Update LICENSE.txt
Change to the AGPL.

Mostly due to the fact that this is a network service, though it's not quite as obvious a decision as if I were writing a webserver or something.

If anyone has any reasons why I should not use the AGPL for Zif, let me know. I'm open to suggestions/opinions.